core/translate: break emitter.rs into smaller modules
