+++ core/translate/expr.rs
@@ -4,14 +4,14 @@ use sqlite3_parser::ast::{self, UnaryOperator};
 use crate::ext::{ExtFunc, UuidFunc};
 #[cfg(feature = "json")]
 use crate::function::JsonFunc;
-use crate::function::{AggFunc, Func, FuncCtx, MathFuncArity, ScalarFunc};
+use crate::function::{Func, FuncCtx, MathFuncArity, ScalarFunc};
 use crate::schema::Type;
 use crate::util::normalize_ident;
 use crate::vdbe::{builder::ProgramBuilder, insn::Insn, BranchOffset};
 use crate::Result;
 
 use super::emitter::Resolver;
-use super::plan::{Aggregate, TableReference, TableReferenceType};
+use super::plan::{TableReference, TableReferenceType};
 
 #[derive(Default, Debug, Clone, Copy)]
 pub struct ConditionMetadata {
@@ -2047,366 +2047,6 @@ pub fn maybe_apply_affinity(col_type: Type, target_register: usize, program: &mu
             register: target_register,
         })
     }
-}
-
-pub fn translate_aggregation(
-    program: &mut ProgramBuilder,
-    referenced_tables: &[TableReference],
-    agg: &Aggregate,
-    target_register: usize,
-    resolver: &Resolver,
-) -> Result<usize> {
-    let dest = match agg.func {
-        AggFunc::Avg => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("avg bad number of arguments");
-            }
-            let expr = &agg.args[0];
-            let expr_reg = program.alloc_register();
-            let _ = translate_expr(program, Some(referenced_tables), expr, expr_reg, resolver)?;
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Avg,
-            });
-            target_register
-        }
-        AggFunc::Count => {
-            let expr_reg = if agg.args.is_empty() {
-                program.alloc_register()
-            } else {
-                let expr = &agg.args[0];
-                let expr_reg = program.alloc_register();
-                let _ = translate_expr(program, Some(referenced_tables), expr, expr_reg, resolver)?;
-                expr_reg
-            };
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Count,
-            });
-            target_register
-        }
-        AggFunc::GroupConcat => {
-            if agg.args.len() != 1 && agg.args.len() != 2 {
-                crate::bail_parse_error!("group_concat bad number of arguments");
-            }
-
-            let expr_reg = program.alloc_register();
-            let delimiter_reg = program.alloc_register();
-
-            let expr = &agg.args[0];
-            let delimiter_expr: ast::Expr;
-
-            if agg.args.len() == 2 {
-                match &agg.args[1] {
-                    ast::Expr::Column { .. } => {
-                        delimiter_expr = agg.args[1].clone();
-                    }
-                    ast::Expr::Literal(ast::Literal::String(s)) => {
-                        delimiter_expr = ast::Expr::Literal(ast::Literal::String(s.to_string()));
-                    }
-                    _ => crate::bail_parse_error!("Incorrect delimiter parameter"),
-                };
-            } else {
-                delimiter_expr = ast::Expr::Literal(ast::Literal::String(String::from("\",\"")));
-            }
-
-            translate_expr(program, Some(referenced_tables), expr, expr_reg, resolver)?;
-            translate_expr(
-                program,
-                Some(referenced_tables),
-                &delimiter_expr,
-                delimiter_reg,
-                resolver,
-            )?;
-
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: delimiter_reg,
-                func: AggFunc::GroupConcat,
-            });
-
-            target_register
-        }
-        AggFunc::Max => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("max bad number of arguments");
-            }
-            let expr = &agg.args[0];
-            let expr_reg = program.alloc_register();
-            let _ = translate_expr(program, Some(referenced_tables), expr, expr_reg, resolver)?;
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Max,
-            });
-            target_register
-        }
-        AggFunc::Min => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("min bad number of arguments");
-            }
-            let expr = &agg.args[0];
-            let expr_reg = program.alloc_register();
-            let _ = translate_expr(program, Some(referenced_tables), expr, expr_reg, resolver)?;
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Min,
-            });
-            target_register
-        }
-        AggFunc::StringAgg => {
-            if agg.args.len() != 2 {
-                crate::bail_parse_error!("string_agg bad number of arguments");
-            }
-
-            let expr_reg = program.alloc_register();
-            let delimiter_reg = program.alloc_register();
-
-            let expr = &agg.args[0];
-            let delimiter_expr = match &agg.args[1] {
-                ast::Expr::Column { .. } => agg.args[1].clone(),
-                ast::Expr::Literal(ast::Literal::String(s)) => {
-                    ast::Expr::Literal(ast::Literal::String(s.to_string()))
-                }
-                _ => crate::bail_parse_error!("Incorrect delimiter parameter"),
-            };
-
-            translate_expr(program, Some(referenced_tables), expr, expr_reg, resolver)?;
-            translate_expr(
-                program,
-                Some(referenced_tables),
-                &delimiter_expr,
-                delimiter_reg,
-                resolver,
-            )?;
-
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: delimiter_reg,
-                func: AggFunc::StringAgg,
-            });
-
-            target_register
-        }
-        AggFunc::Sum => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("sum bad number of arguments");
-            }
-            let expr = &agg.args[0];
-            let expr_reg = program.alloc_register();
-            let _ = translate_expr(program, Some(referenced_tables), expr, expr_reg, resolver)?;
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Sum,
-            });
-            target_register
-        }
-        AggFunc::Total => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("total bad number of arguments");
-            }
-            let expr = &agg.args[0];
-            let expr_reg = program.alloc_register();
-            let _ = translate_expr(program, Some(referenced_tables), expr, expr_reg, resolver)?;
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Total,
-            });
-            target_register
-        }
-    };
-    Ok(dest)
-}
-
-pub fn translate_aggregation_groupby(
-    program: &mut ProgramBuilder,
-    referenced_tables: &[TableReference],
-    group_by_sorter_cursor_id: usize,
-    cursor_index: usize,
-    agg: &Aggregate,
-    target_register: usize,
-    resolver: &Resolver,
-) -> Result<usize> {
-    let emit_column = |program: &mut ProgramBuilder, expr_reg: usize| {
-        program.emit_insn(Insn::Column {
-            cursor_id: group_by_sorter_cursor_id,
-            column: cursor_index,
-            dest: expr_reg,
-        });
-    };
-    let dest = match agg.func {
-        AggFunc::Avg => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("avg bad number of arguments");
-            }
-            let expr_reg = program.alloc_register();
-            emit_column(program, expr_reg);
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Avg,
-            });
-            target_register
-        }
-        AggFunc::Count => {
-            let expr_reg = program.alloc_register();
-            emit_column(program, expr_reg);
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Count,
-            });
-            target_register
-        }
-        AggFunc::GroupConcat => {
-            if agg.args.len() != 1 && agg.args.len() != 2 {
-                crate::bail_parse_error!("group_concat bad number of arguments");
-            }
-
-            let expr_reg = program.alloc_register();
-            let delimiter_reg = program.alloc_register();
-
-            let delimiter_expr: ast::Expr;
-
-            if agg.args.len() == 2 {
-                match &agg.args[1] {
-                    ast::Expr::Column { .. } => {
-                        delimiter_expr = agg.args[1].clone();
-                    }
-                    ast::Expr::Literal(ast::Literal::String(s)) => {
-                        delimiter_expr = ast::Expr::Literal(ast::Literal::String(s.to_string()));
-                    }
-                    _ => crate::bail_parse_error!("Incorrect delimiter parameter"),
-                };
-            } else {
-                delimiter_expr = ast::Expr::Literal(ast::Literal::String(String::from("\",\"")));
-            }
-
-            emit_column(program, expr_reg);
-            translate_expr(
-                program,
-                Some(referenced_tables),
-                &delimiter_expr,
-                delimiter_reg,
-                resolver,
-            )?;
-
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: delimiter_reg,
-                func: AggFunc::GroupConcat,
-            });
-
-            target_register
-        }
-        AggFunc::Max => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("max bad number of arguments");
-            }
-            let expr_reg = program.alloc_register();
-            emit_column(program, expr_reg);
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Max,
-            });
-            target_register
-        }
-        AggFunc::Min => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("min bad number of arguments");
-            }
-            let expr_reg = program.alloc_register();
-            emit_column(program, expr_reg);
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Min,
-            });
-            target_register
-        }
-        AggFunc::StringAgg => {
-            if agg.args.len() != 2 {
-                crate::bail_parse_error!("string_agg bad number of arguments");
-            }
-
-            let expr_reg = program.alloc_register();
-            let delimiter_reg = program.alloc_register();
-
-            let delimiter_expr = match &agg.args[1] {
-                ast::Expr::Column { .. } => agg.args[1].clone(),
-                ast::Expr::Literal(ast::Literal::String(s)) => {
-                    ast::Expr::Literal(ast::Literal::String(s.to_string()))
-                }
-                _ => crate::bail_parse_error!("Incorrect delimiter parameter"),
-            };
-
-            emit_column(program, expr_reg);
-            translate_expr(
-                program,
-                Some(referenced_tables),
-                &delimiter_expr,
-                delimiter_reg,
-                resolver,
-            )?;
-
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: delimiter_reg,
-                func: AggFunc::StringAgg,
-            });
-
-            target_register
-        }
-        AggFunc::Sum => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("sum bad number of arguments");
-            }
-            let expr_reg = program.alloc_register();
-            emit_column(program, expr_reg);
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Sum,
-            });
-            target_register
-        }
-        AggFunc::Total => {
-            if agg.args.len() != 1 {
-                crate::bail_parse_error!("total bad number of arguments");
-            }
-            let expr_reg = program.alloc_register();
-            emit_column(program, expr_reg);
-            program.emit_insn(Insn::AggStep {
-                acc_reg: target_register,
-                col: expr_reg,
-                delimiter: 0,
-                func: AggFunc::Total,
-            });
-            target_register
-        }
-    };
-    Ok(dest)
 }
 
 /// Get an appropriate name for an expression.
